pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pip-toolkit"
version = "2.0.0"
description = "PrivEsc Intelligence Platform — verified, ranked Linux privilege escalation paths."
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"
keywords = ["penetration-testing", "privilege-escalation", "security", "linux", "redteam"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Information Technology",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
]
dependencies = [
    "typer[all]>=0.12.0",
    "rich>=13.7.0",
    "pydantic>=2.6.0",
    "networkx>=3.3",
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "jinja2>=3.1.4",
]

[project.optional-dependencies]
pdf = ["weasyprint>=62.0"]
dev = [
    "pytest>=8.1.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
    "bandit>=1.7.8",
]

[project.scripts]
pip-toolkit = "pip:app"

[project.urls]
Homepage   = "https://github.com/yourusername/pip-toolkit"
Repository = "https://github.com/yourusername/pip-toolkit"
Issues     = "https://github.com/yourusername/pip-toolkit/issues"


# ── pytest ─────────────────────────────────────────────────────────────────────
[tool.pytest.ini_options]
testpaths       = ["tests"]
asyncio_mode    = "auto"
addopts         = "-v --tb=short --cov=pip --cov-report=term-missing"
filterwarnings  = ["ignore::DeprecationWarning"]


# ── ruff (linter + formatter) ──────────────────────────────────────────────────
[tool.ruff]
target-version = "py311"
line-length    = 100

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "S",   # bandit (security)
    "N",   # pep8-naming
]
ignore = [
    "S603",  # subprocess calls — intentional in a pentest tool
    "S604",  # shell=True — intentional
    "S605",  # os.system — intentional
    "S607",  # partial executable path — intentional
    "S101",  # assert — used in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/*"  = ["S", "N"]   # Relax security and naming rules in tests
"plugins/*" = ["N"]        # Plugins may use different naming conventions


# ── mypy ──────────────────────────────────────────────────────────────────────
[tool.mypy]
python_version          = "3.11"
strict                  = false
ignore_missing_imports  = true
warn_return_any         = true
warn_unused_ignores     = true
exclude                 = ["plugins/", "tests/"]


# ── coverage ──────────────────────────────────────────────────────────────────
[tool.coverage.run]
source   = ["pip"]
omit     = ["pip/api/*", "tests/*", "plugins/*"]

[tool.coverage.report]
show_missing  = true
skip_covered  = false
fail_under    = 60
